agent-server/requirements.txt
fastapi
uvicorn
langchain
requests
beautifulsoup4
playwright
python-dotenv
